CoreCommands: make `SwiftToolObservabilityHandler` public
This allows `ObservabilityScope` to be created with this handler for logging purposes.
